requirements.txt
Flask==2.0.1
requests==2.31.0
python-dotenv==0.19.0
PyPDF2==1.26.0
python-docx==0.8.11
openpyxl==3.0.9
pytesseract==0.3.8
pdf2image==1.16.0
werkzeug==2.0.3
reportlab
pdfplumber
